# [docs] Improve Link component