--- a/docs/src/components/Link.tsx
+++ b/docs/src/components/Link.tsx
@@ -3,19 +3,22 @@ import clsx from 'clsx';
 import NextLink from 'next/link';
 import { ExternalLinkIcon } from 'docs/src/icons/ExternalLinkIcon';
 
-export function Link({ href, className, ...props }: React.ComponentProps<typeof NextLink>) {
+export function Link(props: React.ComponentProps<typeof NextLink>) {
+  const { href, className } = props;
+  let pathname = typeof href === 'string' ? href : href.pathname!;
+
   // Sometimes link come from component descriptions; in this case, remove the domain
-  if (typeof href === 'string' && href.startsWith('https://base-ui.com')) {
-    href = href.replace('https://base-ui.com', '');
+  if (pathname.startsWith('https://base-ui.com/')) {
+    pathname = pathname.replace('https://base-ui.com/', '/');
   }
 
-  if (typeof href === 'string' && href.startsWith('http')) {
+  if (pathname.startsWith('http')) {
     return (
       <NextLink
-        href={href}
         target="_blank"
         rel="noopener"
         {...props}
+        href={href}
         className={clsx('Link mr-[0.125em] inline-flex items-center gap-[0.25em]', className)}
       >
         {props.children}
@@ -24,10 +27,10 @@ export function Link({ href, className, ...props }: React.ComponentProps<typeof 
     );
   }
 
-  if (typeof href === 'string' && (href.endsWith('.md') || href.endsWith('.txt'))) {
-    // Relative url, but outside the Next.js router
-    return <a href={href} className={clsx('Link', className)} {...props} />;
+  if (pathname.endsWith('.md') || pathname.endsWith('.txt')) {
+    // Relative URL, but outside the Next.js router
+    return <a {...props} href={pathname} className={clsx('Link', className)} />;
   }
 
-  return <NextLink href={href} className={clsx('Link', className)} {...props} />;
+  return <NextLink {...props} className={clsx('Link', className)} />;
 }