A few fixes for fun:

1. There is https://github.com/mui/base-ui/security/code-scanning/20. Fixed
2. The href type https://nextjs.org/docs/pages/api-reference/components/link#reference can be an object too. Fixed
3. Avoid destructuring objects when possible (using ...)
4. We have a heavy DOM structure for nothing; we don't need those 42 `SideNavLabel` span nodes. The more components we add, the heavier:

<img width="302" height="113" alt="SCR-20250816-peyz" src="https://github.com/user-attachments/assets/383c9130-47bb-4235-8cb4-2a988585a2c7" />

Fixed

5. By default, Next.js prefetches all the links as soon as they enter the viewport (https://nextjs.org/docs/app/api-reference/components/link#prefetch). For a side nav, it spams requests to the CDN for nothing:

https://github.com/user-attachments/assets/906396b3-7bbf-44a0-84c9-7f4db10bb91e

Instead, wait for the user to hover link. Fixed

6. Avoiding binding onClick so many times. Fixed

Preview: https://deploy-preview-2519--base-ui.netlify.app/react/overview/quick-start