@colmtuite 2025-08-20
Also, `onMouseEnter:` only will break prefetching on focus. So we'd need to add prefetching for focus too. But it seems it's much better to rely on next/link and prefetch when the link enters the viewport