@oliviertassinari 2025-08-27
> Testing this, the sidebar links feel much slower? On prod, 

@colmtuite Production has a better hosting infrastructure than the Netlify previews. To compare the experience, I think we should take another preview.

- This PR: https://deploy-preview-2519--base-ui.netlify.app/react/overview/quick-start
- Baseline: https://deploy-preview-2515--base-ui.netlify.app/react/overview/quick-start

> you need to hover the link for a bit before clicking it, to ensure the page loads in time?

Right, so when I record myself, it takes me 250ms from reaching the link that I want to navigate to and pressing `mousedown`. It then takes me another 150ms to release my pointer to trigger `mouseup` at which point the link navigation starts. So the time budget is 400ms from hover to navigation. I guess I'm slow, so people who are faster than me, would take at the minimum 250ms? 

<img width="1193" height="460" alt="SCR-20250828-cebb" src="https://github.com/user-attachments/assets/f4b6548d-7978-418d-b77c-d79266102c8e" />

One given page takes <100ms to load, so it's not an issue for decent connetions. I think the problem is more than if I hover a bunch of pages, say 10, I can starve the network, by the time I want to click on the last menu item, the page still loads the JS of the links that are hidden, and it will feel slow.

I can kind of reproduce the issue you described (sometimes). The benefit of only pre-fetching on hover is to not have TB of TB of docs traffic, but considering that AI pressures this kind of traffic down (I don't expect the bots going to our website to have JS enabled), if you still feel that it's slow, let's give up on this. Bandwidth should be fine; let's prioritize UX.

> onMouseEnter: only will break prefetching on focus.

I assumed that we don't care 😆