@colmtuite 2025-08-28
> Bandwidth is not that big of a concern, not if we can improve the UX

But this is an enormous downgrade in UX. I've attached a video of me testing the baseline link you provided with production. The use case is I want to check the Toast docs, so I open the website, scroll to "T", find Toast and click it. You can see the PR is significantly slower.

https://github.com/user-attachments/assets/06a79f1b-2c75-418b-a122-47e47b621912

> I guess I'm slow, so people who are faster than me, would take at the minimum 250ms

For me and many others, these interactions are 50ms–100ms. Also, connections can be weak, even in good environments.
 
Side note: overriding core dependency features like this is generally a bad idea imo. It breaks updates. So if the Next.js team improve something in next/link, this custom code might break the update. There's no need to spend time optimising things like this right now. It's working fine, nobody is complaining about it, and it sucks time from much more important work. So I'll close this.

cc @oliviertassinari