@oliviertassinari 2025-08-28
> There's no need to spend time optimising things like this right now. It's working fine, nobody is complaining about it

@colmtuite Right, so you feel that it's not worth the time, it's not worth risking downgrading the UX. OK, it's fair, let's ignore point 5 (prefech change).

The context: Netlify in the past threatened to downgrade our unlimited free OSS plan because we were using crazy amount of bandwidth. We used to download all the pages of the docs to make them available offline 🙃. We did point 5 in Material UI to keep their good grace. 

I will open a new one with without point 5, the rest are quick wins, only about engineering and it's about correctly using Next.js Links, in it's most default mode. 

---

> You can see the PR is significantly slower

Looking at the video recording, I think we should expect this behavior from any PR. This is comparing the **production infrastructure** of Netlify with the **preview PR** infrastructure of Netlify.

> these interactions are 50ms–100ms

Per https://www.reddit.com/r/pcgaming/comments/iksbbi/so_apparently_150ms_is_the_average_reaction_time/, I would expect the average designer reaction time to be higher, 100ms+ and the time it takes to release the mouse 50ms+. Now, I agree, I could reproduce slowness in one case:

- hovering a bunch of new links all at once
- clicking on the last link